Refactored to remove isSiteEligibleForCustomerHome selector
