+++ client/my-sites/customer-home/main.jsx
@@ -27,11 +27,13 @@ import { preventWidows } from 'lib/formatting';
 import SidebarNavigation from 'my-sites/sidebar-navigation';
 import { SIDEBAR_SECTION_TOOLS } from 'my-sites/sidebar/constants';
 import { getSelectedSite, getSelectedSiteId, getSelectedSiteSlug } from 'state/ui/selectors';
-import { getCustomizerUrl, getSiteOption } from 'state/sites/selectors';
+import {
+	canCurrentUserUseCustomerHome,
+	getSiteFrontPage,
+	getCustomizerUrl,
+	getSiteOption,
+} from 'state/sites/selectors';
 import { getDomainsBySiteId } from 'state/sites/domains/selectors';
-import getSiteFrontPage from 'state/sites/selectors/get-site-front-page';
-import canCurrentUserUseCustomerHome from 'state/sites/selectors/can-current-user-use-customer-home';
-import isSiteEligibleForCustomerHome from 'state/selectors/is-site-eligible-for-customer-home';
 import PageViewTracker from 'lib/analytics/page-view-tracker';
 import DocumentHead from 'components/data/document-head';
 import getSiteChecklist from 'state/selectors/get-site-checklist';
@@ -80,7 +82,6 @@ class Home extends Component {
 				);
 			}
 		},
-		isSiteEligible: PropTypes.bool.isRequired,
 		expandToolsAndTrack: PropTypes.func.isRequired,
 		trackAction: PropTypes.func.isRequired,
 		isStaticHomePage: PropTypes.bool.isRequired,
@@ -121,9 +122,7 @@ class Home extends Component {
 		const { translate, canUserUseCustomerHome, siteSlug } = this.props;
 
 		if ( ! canUserUseCustomerHome ) {
-			const title = this.props.isSiteEligible
-				? translate( 'You are not authorized to view this page.' )
-				: translate( 'This page is not available on this site.' );
+			const title = translate( 'This page is not available on this site.' );
 			return (
 				<EmptyContent
 					title={ preventWidows( title ) }
@@ -403,7 +402,6 @@ const connectHome = connect(
 			canUserUseCustomerHome: canCurrentUserUseCustomerHome( state, siteId ),
 			hasChecklistData,
 			isChecklistComplete,
-			isSiteEligible: isSiteEligibleForCustomerHome( state, siteId ),
 			isStaticHomePage: 'page' === getSiteOption( state, siteId, 'show_on_front' ),
 			staticHomePageId: getSiteFrontPage( state, siteId ),
 			showCustomizer: ! isSiteUsingFullSiteEditing( state, siteId ),
